Change GSR domain model type from Int to Float

+++ app/src/main/java/pl/edu/pjwstk/engineeringthesis/model/GsrSample.kt
@@ -4,5 +4,5 @@ data class GsrSample(
     val id : Int,
     val userId : Int,
     val epoch : Long,
-    val gsr : Int
+    val gsr : Float
 )
